python_app/requirements.txt
psycopg2-binary
paho-mqtt
psutil